Move most functionality from libbeat/beat to libbeat/cmd package
- only have commonly shared structs/interfaces in libbeat/beat
- move libbeat/beat functionality to libbeat/cmd/instance package
- update imports/dependencies
- start hiding some exported symbols
