+++ libbeat/beat/error.go
@@ -1,0 +1,6 @@
+package beat
+
+import "errors"
+
+// GracefulExit is an error that signals to exit with a code of 0.
+var GracefulExit = errors.New("graceful exit")
