+++ libbeat/cmd/instance/flags.go
@@ -1,4 +1,8 @@
-package beat
+package instance
+
+import "github.com/elastic/beats/libbeat/beat"
+
+// XXX: With introduction of cmd package, the FlagsHandler should be deprecated.
 
 // FlagsHandler is an interface that can optionally be implemented by a Beat
 // if it needs to process command line flags on startup. If implemented, the
@@ -6,10 +10,10 @@ package beat
 // and before any of the Beater interface methods are invoked. There will be
 // no callback when '-help' or '-version' are specified.
 type FlagsHandler interface {
-	HandleFlags(*Beat) error // Handle any custom command line arguments.
+	HandleFlags(*beat.Beat) error // Handle any custom command line arguments.
 }
 
-type FlagsHandlerCallback func(*Beat) error
+type FlagsHandlerCallback func(*beat.Beat) error
 
 var handlers []FlagsHandler
 
@@ -17,11 +21,11 @@ func AddFlagsHandler(h FlagsHandler) {
 	handlers = append(handlers, h)
 }
 
-func AddFlagsCallback(cb func(*Beat) error) {
+func AddFlagsCallback(cb func(*beat.Beat) error) {
 	AddFlagsHandler(FlagsHandlerCallback(cb))
 }
 
-func handleFlags(b *Beat) error {
+func handleFlags(b *beat.Beat) error {
 	for _, h := range handlers {
 		err := h.HandleFlags(b)
 		if err != nil {
@@ -31,6 +35,6 @@ func handleFlags(b *Beat) error {
 	return nil
 }
 
-func (cb FlagsHandlerCallback) HandleFlags(b *Beat) error {
+func (cb FlagsHandlerCallback) HandleFlags(b *beat.Beat) error {
 	return cb(b)
 }
