+++ libbeat/cmd/instance/metrics.go
@@ -1,4 +1,4 @@
-package beat
+package instance
 
 import (
 	"runtime"
