+++ libbeat/cmd/run.go
@@ -8,6 +8,7 @@ import (
 	"github.com/spf13/pflag"
 
 	"github.com/elastic/beats/libbeat/beat"
+	"github.com/elastic/beats/libbeat/cmd/instance"
 )
 
 func genRunCmd(name, version string, beatCreator beat.Creator, runFlags *pflag.FlagSet) *cobra.Command {
@@ -15,7 +16,7 @@ func genRunCmd(name, version string, beatCreator beat.Creator, runFlags *pflag.F
 		Use:   "run",
 		Short: "Run " + name,
 		Run: func(cmd *cobra.Command, args []string) {
-			err := beat.Run(name, version, beatCreator)
+			err := instance.Run(name, version, beatCreator)
 			if err != nil {
 				os.Exit(1)
 			}
