+++ libbeat/cmd/test/config.go
@@ -7,6 +7,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/elastic/beats/libbeat/beat"
+	"github.com/elastic/beats/libbeat/cmd/instance"
 )
 
 func GenTestConfigCmd(name, version string, beatCreator beat.Creator) *cobra.Command {
@@ -14,7 +15,7 @@ func GenTestConfigCmd(name, version string, beatCreator beat.Creator) *cobra.Com
 		Use:   "config",
 		Short: "Test configuration settings",
 		Run: func(cmd *cobra.Command, args []string) {
-			b, err := beat.New(name, version)
+			b, err := instance.NewBeat(name, version)
 			if err != nil {
 				fmt.Fprintf(os.Stderr, "Error initializing beat: %s\n", err)
 				os.Exit(1)
