+++ libbeat/cmd/test/output.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/elastic/beats/libbeat/beat"
+	"github.com/elastic/beats/libbeat/cmd/instance"
 	"github.com/elastic/beats/libbeat/outputs"
 	"github.com/elastic/beats/libbeat/testing"
 )
@@ -16,7 +16,7 @@ func GenTestOutputCmd(name, beatVersion string) *cobra.Command {
 		Use:   "output",
 		Short: "Test output works with current settings",
 		Run: func(cmd *cobra.Command, args []string) {
-			b, err := beat.New(name, beatVersion)
+			b, err := instance.NewBeat(name, beatVersion)
 			if err != nil {
 				fmt.Fprintf(os.Stderr, "Error initializing beat: %s\n", err)
 				os.Exit(1)
